changes for Python 3 compatibility
Added decoding of compiler flags to UTF-8. Building using Python 3 (on Mac) didn't include the include directories (although they were found in setup.py) because the string comparison was failing for bytes literals. Also moved the c++ language option to the Cython files, as the language option for cythonize is deprecated.